Refactor Parser again
Have the Parser create the Lexer; there's no point in passing a Lexer
into the Parser when the Parser can just create the Lexer itself.

Also revert the token handling - have the Parser use the Lexer to
tokenize the entire input before starting, as we'll need the ability to
"rewind" the parser if we make an assumption about what kind of block we
are parsing and end up being wrong.