ALMOST HAVE SYNTAX HIGHLIGHTING!!
Currently have the ability to detect the language (from the code-block
tag) and load the TreeSitter parser and highligh query, then run the
query against the code block.

Last step: Defining a standard highlighting scheme (match capture groups
to colors), create a data structure to tag ranges of the source with the
capture group, then loop through and print each range in the assigned
color.