html, serve: Allow CSS override from command line
The 'serve' command now takes an optional 'css' subcommand where all
of the various styles (e.g. heading-1, code-block, etc.) can be
overridden.

The 'flags' dependency was updated to accomplish this: The Css struct
itself is used as the command, so as fields are added / removed to
that struct, both the render code and the CLI will stay in sync.