lua: Fix load time
Previously, the 'setup()' function would _also_ check if Zigdown was
already built, and if not, build it. Doing all of the setup work
necessary to start the background build job added a noticeable delay to
Neovim's startup.

Now, the 'setup()' function only assigns config options; building and
loading of the plugin is deferred to the first render or format call.