Refactor command-line args for all renderers
Each renderer has slightly different options to configure the output, so
each now has its own unique options struct.

Because we use `render.zig` to handle the implementation of the
rendering outside of `main.zig`, we now have `cli.zig` to share the
common options structs to be used from `main()` (for use with Flags) and
from `render()` (to map to the renderer-specific config structs).

Along with this refactor, the HtmlRenderer now supports generating only
the body of the HTML document (excludes the `<html><body>...</html></body>`
tags), and the ConsoleRenderer can be told to not fetch non-local images
from the internet.