Tokenizer: Handle UTF-8 codepoints
Previously, all tokens were assumed to contain only ASCII characters,
resulting in incorrect token parsing when a non-first byte of a multi-byte
UTF-8 codepoint was a valid ASCII value corresponding to a scalar token
such as '>' or '*'.

To fix this, the first Tokenizer used by the Lexer checks if next
character is a multi-byte UTF-8 codepoint, and if so, treats the
following bytes as a Word token that may contain non-ASCII characters.

The WordTokenizer handles this by performing the same check on each byte
it encounters, skipping the necessary number of bytes to glob the
current codepoint into the surrounding Word token.