utils: Add generateNavigationTree
Given a root file (and its directory), recursively parse each Markdown
document, scan for links to other Markdown files, and follow them.

The idea is to generate a navigation tree when serving up documents in
HTML form.

The next step will be to take this tree data structure and generate HTML
from it within the `html` and `serve` commands of Zigdown.