parser: Add table alignment and relative width parsing
Table columns are left-aligned by default, but `:--:` or `--:` will
specify center or right alignment, respectively.

Column widths are specified by the number of dashes `-` used in each
column. The final width of the table will still be the entire width of
the document being rendered to, but each column will receive a width
proportional to the fraction of `-` characters compared to the total
number. At least, that is the current plan.

This just implements the parsing (the easy part); the rendering will be
more difficult and take some time to fully implement.