build.zig: Fix ziglua lazy-dependency issue
It appears that the Zig build system has a bug where it will not fetch
lazy dependencies of lazy dependencies. For Zigdown, this means that in
order to build the Lua module using ziglua, we must manually reference
the necessary luajit lazy-dependency of ziglua so that it can build
luajit for us.